scm: Fix alert for possible overflow
lgtm.com shows the following alert:

	Multiplication result may overflow 'unsigned int' before it is
	converted to 'size_t'.

+++ include/common/scm-code.c
@@ -3,7 +3,7 @@
 #endif
 
 static void scm_fdset_init_chunk(struct scm_fdset *fdset, int nr_fds,
-		void *data, unsigned ch_size)
+		void *data, size_t ch_size)
 {
 	struct cmsghdr *cmsg;
 	static char dummy;
